Identifier/requirements.txt
fastapi~=0.103.0
uvicorn~=0.20.0
python-dotenv~=0.21.0
openai~=1.25.0
pydantic~=2.5.3
